// node/modulo-1/aula06-post/atividade_avaliativa/src/enums/messages.ts
enum messages {
    clientNotFind = "Cliente não encontrado",
    clientNotSet = "Cliente não cadastrado!",
    clientRegistrerSuccess = "Cliente cadastrado com sucesso!",
    clientUpdatedSuccess = "Cliente atualizado com sucesso!",
    clientRemoveSuccess = "Cliente removido com sucesso!",
    clientNotUpdated = "Cliente não atualizado",
  }
  
  export default messages;